Skip ActiveJob retrying for ProcessInstallmentJob
This disables the ActiveJob default retry mechanism because installment
processing failures are already taken care of by solidus_subscriptions.

Before this, there was a potentional for double retrying, which could
causes background jobs and cancelled orders to accumulate leading
to side effects such as a bad customer experience.

Now, a `process_job_error_handler` configuration Proc is called when a
ProcessInstallmentJob `#perfom` method fails.
The rescued error can be managed/logged as preffered, or re-raised
if the default retry behaviour is required in the job.